Fix MQTT switch off message and clean up const usage
- Fix Machine::power_mqtt sending the "on" message even when switching off
- Add virtual destructor to polymorphic MQTTInterface::Response base class
- Remove const qualifier from by-value return types across the codebase
- Replace using namespace std::chrono_literals with explicit operator usings
- Use size_t instead of int for CachedCards indices
- Return Buzzer& instead of Buzzer* from getBuzzerForTesting
- Remove unused extern FabBackend declaration in AuthProvider.cpp
- Remove redundant copy assignment in MachineConfig, make Task instances non-const
- Recommend pioarduino VS Code extension

+++ include/CachedCards.hpp
@@ -28,7 +28,7 @@ namespace fabomatic
 
     constexpr CachedCards() : cards{card::INVALID}, levels{FabUser::UserLevel::Unknown} {};
 
-    constexpr auto operator[](int i) const -> const CachedCard
+    constexpr auto operator[](size_t i) const -> const CachedCard
     {
       return {cards[i], levels[i]};
     }
@@ -48,7 +48,7 @@ namespace fabomatic
       return std::nullopt;
     }
 
-    constexpr auto set_at(int idx, const card::uid_t &uid, const FabUser::UserLevel &level) -> void
+    constexpr auto set_at(size_t idx, const card::uid_t &uid, const FabUser::UserLevel &level) -> void
     {
       cards[idx] = uid;
       levels[idx] = level;
